switch GLOBAL to LOCAL added
